Use searchkit builder to show line-item records in contribution form

See merge request extensions/lineitemedit!83

+++ CRM/Lineitemedit/Util.php
@@ -504,7 +504,7 @@ ORDER BY  ps.id, pf.weight ;
         )
       );
       $contriParams['total_amount'] = $updatedAmount;
-      $contriParams['net_amount'] = $updatedAmount - ($updatedContribution['fee_amount'] ?? 0);
+      $contriParams['net_amount'] = (float) $updatedAmount - (float) ($updatedContribution['fee_amount'] ?? 0);
       if ($taxAmount) {
         $contriParams['tax_amount'] = $taxAmount;
       }
